T8-CLANG-TIDY-BASELINE-CLEANUP: clear clang-tidy allowed-red baseline carried since PR #56
Mechanical (option a) batch-fix across the 13 src/ .cpp files and one
header (audio_file.hpp, auto-fix-driven signature change matched to
its existing move-init body) to bring the clang-tidy CI job from fail
to pass for the first time since the Tier 7 cycle opened at PR #56
(2026-05-19).

Cure technique mix:
  * clang-tidy --fix --fix-errors cleared the four high-volume mechanical
    classes (braces-around-statements 72, uppercase-literal-suffix 57,
    return-braced-init-list 42, use-auto 40) and several long-tail classes
    (use-ranges, use-designated-initializers, implicit-bool-conversion,
    math-missing-parentheses, use-integer-sign-comparison, init-variables,
    use-emplace, pass-by-value).
  * Hand-corrections at 3 sites where the use-integer-sign-comparison
    auto-fix mangled paren balance around a static_cast (reference_mode,
    app, app_handlers).
  * C-array -> std::array conversion for audio_file.cpp's 11 magic-byte
    arrays, with compare_bytes retyped to take std::span<const uint8_t>
    (drops explicit len at 17 call sites). Also float80[10] -> std::array
    in build_aiff_header and `static const char hex[]` -> std::string_view
    in reaper_export.
  * misplaced-widening-cast: cast moved inside the addition at 2 sites.
  * implicit-widening-of-multiplication-result: kTailWindow widened at
    the first factor.
  * performance-enum-size: local AudioFormat retyped to : std::uint8_t.
  * avoid-nested-conditional-operator: nested ternary expanded to ifs.
  * deadcode.DeadStores: removed dead fade_end_s assignment surfaced once
    the high-volume mechanical noise was cleared.

NOLINT annotations added (15 total):
  * reinterpret-cast 5 sites — binary IO (fstream::read/write on
    uint8_t/byte buffers) + intentional pointer-bits-as-entropy for
    RNG seed; all load-bearing per Core Guidelines exception.
  * function-cognitive-complexity 4 functions (parse_wav_header, main,
    align_per_track, run_tui) — pipeline/wiring; decomposition tracked
    separately.
  * exception-escape 1 site (main) — paired with cog-complexity block.
  * branch-clone 3 NOLINTs covering 4 logical sites — explicit
    default-to-WAV fallback and placeholder TUI rendering branches.
  * return-braced-init-list 1 site (analysis.cpp) — `{count, value}`
    would be parsed as 2-element initializer_list<float>, not the
    (size, value) vector ctor.

src/core/test_deps.cpp deleted (Mi-11): file was not referenced by
CMakeLists.txt, not compiled into any target, and contained only
header-availability probes. Removed clang-tidy's clang-diagnostic-error
that the file produced because it had no compile DB entry. Mi-11 also
moves to RESOLVED in BACKLOG.md.

Behavior preservation: ctest 14/14 pass on build-clang-tidy/ Debug
config, pre-cure and post-cure, with no test modifications. The
existing test suite is the regression-guard per mandate.

Final post-cure local clang-tidy on src/**/*.cpp excluding vendored
pocketfft_hdronly.h (per CI workflow's `-not -path '*/pocketfft*'`):
0 errors.

Local-only pre-existing noise: Brew Homebrew LLVM 20.1.4 surfaces 5
clang-analyzer-cplusplus.NewDeleteLeaks warnings against pocketfft
header that appear in both pre-cure and post-cure logs. They are
NOT in the CI clang-tidy baseline; Ubuntu CI's older clang-tidy
does not surface them and the mandate forbids modifying pocketfft.

+++ src/main.cpp
@@ -66,6 +66,8 @@ void print_help() {
               << "  -h, --help               Show this help\n";
 }
 
+// NOLINTBEGIN(bugprone-exception-escape, readability-function-cognitive-complexity)
+// — argument-parsing main; refactor into smaller helpers tracked separately.
 int main(int argc, char* argv[]) {
     if (argc < 2) {
         print_help();
@@ -171,7 +173,7 @@ int main(int argc, char* argv[]) {
                       << std::setw(2) << start_min << ":" << std::setw(2) << start_s << " - "
                       << std::setw(2) << end_min << ":" << std::setw(2) << end_s
                       << " (" << static_cast<int>(duration / 60) << "m " 
-                      << static_cast<int>(static_cast<int>(duration) % 60) << "s)"
+                      << (static_cast<int>(duration) % 60) << "s)"
                       << " — confidence: " << std::fixed << std::setprecision(2) << sp.confidence
                       << "\n";
         }
@@ -322,7 +324,7 @@ int main(int argc, char* argv[]) {
                       << std::setw(2) << start_min << ":" << std::setw(2) << start_s << " - "
                       << std::setw(2) << end_min << ":" << std::setw(2) << end_s
                       << " (" << static_cast<int>(duration / 60) << "m " 
-                      << static_cast<int>(static_cast<int>(duration) % 60) << "s)"
+                      << (static_cast<int>(duration) % 60) << "s)"
                       << " — confidence: " << std::fixed << std::setprecision(2) << sp.confidence
                       << "\n";
         }
@@ -402,3 +404,4 @@ int main(int argc, char* argv[]) {
     print_help();
     return 1;
 }
+// NOLINTEND(bugprone-exception-escape, readability-function-cognitive-complexity)
